Move load_schema from Object to AR::TestCase
as its instance and class mathod
