+++ activerecord/test/cases/test_case.rb
@@ -11,6 +11,7 @@ require "cases/validations_repair_helper"
 require_relative "../support/config"
 require_relative "../support/connection"
 require_relative "../support/adapter_helper"
+require_relative "../support/load_schema_helper"
 
 module ActiveRecord
   # = Active Record Test Case
@@ -23,6 +24,8 @@ module ActiveRecord
     include ActiveRecord::ValidationsRepairHelper
     include AdapterHelper
     extend AdapterHelper
+    include LoadSchemaHelper
+    extend LoadSchemaHelper
 
     self.fixture_path = FIXTURES_ROOT
     self.use_instantiated_fixtures = false
@@ -225,6 +228,8 @@ module ActiveRecord
 
     # Connect to the database
     ARTest.connect
+    # Load database schema
+    load_schema
   end
 
   class PostgreSQLTestCase < TestCase
